feat: show apps with available updates
